# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: IncentEdge API
  version: 1.0.0
  description: |
    # IncentEdge API Documentation

    IncentEdge is an AI-powered incentive discovery, application, and monetization platform - Infrastructure's Bloomberg Terminal for Incentives.

    ## Features
    - **24,458+ verified incentive programs** across federal, state, local, and utility levels
    - **AI-powered matching** using tiered LLM system (DeepSeek + Claude)
    - **Direct Pay eligibility checking** for IRA Section 6417
    - **Multi-format report generation** (PDF, HTML, JSON)
    - **Real-time analytics** and portfolio tracking

    ## Authentication
    All authenticated endpoints require either:
    - Session-based auth (via Supabase Auth cookies)
    - API Key authentication (via `X-API-Key` header)

    ## Rate Limiting
    Rate limits vary by subscription tier:
    - **Free**: 100 requests/hour
    - **Starter**: 1,000 requests/hour
    - **Professional**: 10,000 requests/hour
    - **Team**: 50,000 requests/hour
    - **Enterprise**: Custom limits

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Error Handling
    The API uses standard HTTP response codes:
    - `200` - Success
    - `201` - Created
    - `400` - Bad Request (validation errors)
...
    - `404` - Not Found
...

...
    ```json
    {
...
servers:
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session token (automatic via cookies)

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
...
paths:
...
  /applications/{id}/submit:
    post:
      summary: Submit Application
      description: Submit application to the program administrator
      tags:
        - Applications
      security:
        - SessionAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application submitted successfully
...